+++ telegram/message_service.py
@@ -1,70 +1,100 @@
-from common.utils import compose
-from telegram import telegram, telegram_repo, payload_repo
+import yaml
 
-# ----------------------------------- Send ----------------------------------- #
+from netsuite.sales_order.sales_order_repo import get_url
+from telegram import telegram, telegram_repo
+
+DIVIDER = "\=\=\=\=\=\=\=\=\=\=\="
 
 
 def send_new_order(channel: telegram.Channel):
-    def _send(order: dict, prepared_id: str) -> tuple[dict, str]:
+    def _send(order: dict, id: str) -> tuple[dict, str]:
         telegram_repo.send(
-            channel,
-            compose(
-                telegram_repo.build_send_payload(
-                    payload_repo.add_new_order, channel.ecom, order, prepared_id
+            {
+                "chat_id": channel.chat_id,
+                "text": "\n".join(
+                    [
+                        f"Đơn hàng *{channel.ecom}* mới",
+                        DIVIDER,
+                        f"`{id}`",
+                        DIVIDER,
+                        "```",
+                        yaml.dump(order, allow_unicode=True),
+                        "```",
+                    ]
                 ),
-                telegram_repo.build_send_payload(
-                    payload_repo.add_new_order_callback,
-                    prepared_id,
-                ),
-            ),
+                "reply_markup": {
+                    "inline_keyboard": [
+                        [
+                            {
+                                "text": "Tạo đơn",
+                                "callback_data": telegram_repo.build_callback_data(
+                                    "O",
+                                    1,
+                                    id,
+                                ),
+                            },
+                        ]
+                    ],
+                },
+            }
         )
-        return order, prepared_id
+        return order, id
 
     return _send
 
 
-def send_create_order_success(chat_id: str, message_id: int, prepared_id: str):
+def send_create_order_success(chat_id: str, message_id: int, id: str):
     def _send(res: tuple[int, str]) -> tuple[int, str]:
-        id, memo = res
+        _id, memo = res
         telegram_repo.send(
-            telegram.Channel("", chat_id),
-            compose(
-                telegram_repo.build_send_payload(
-                    payload_repo.add_create_order_success,
-                    id,
-                    memo,
+            {
+                "chat_id": chat_id,
+                "reply_to_message_id": message_id,
+                "text": "\n".join(
+                    [
+                        f"Tạo đơn hàng `{memo}` thành công^^",
+                        DIVIDER,
+                        f"Check ngay: [{get_url(str(_id))}]({get_url(str(_id))})",
+                    ]
                 ),
-                telegram_repo.build_send_payload(
-                    payload_repo.add_create_order_callback,
-                    prepared_id,
-                ),
-                telegram_repo.build_send_payload(
-                    payload_repo.add_message_reply,
-                    message_id,
-                ),
-            ),
+                "reply_markup": {
+                    "inline_keyboard": [
+                        [
+                            {
+                                "text": "Đóng đơn",
+                                "callback_data": telegram_repo.build_callback_data(
+                                    "O",
+                                    -1,
+                                    id,
+                                ),
+                            },
+                        ]
+                    ]
+                },
+            }
         )
         return res
 
     return _send
 
 
-def send_create_order_error(chat_id: str, message_id: int, prepared_id: str):
+def send_create_order_error(chat_id: str, message_id: int):
     def _send(res: tuple[Exception, str]) -> tuple[Exception, str]:
         err, memo = res
         telegram_repo.send(
-            telegram.Channel("", chat_id),
-            compose(
-                telegram_repo.build_send_payload(
-                    payload_repo.add_create_order_error,
-                    err,
-                    memo,
+            {
+                "chat_id": chat_id,
+                "reply_to_message_id": message_id,
+                "text": "\n".join(
+                    [
+                        f"Tạo đơn hàng `{memo}` thất bại X\.X",
+                        DIVIDER,
+                        "```",
+                        repr(err),
+                        "```",
+                    ]
                 ),
-                telegram_repo.build_send_payload(
-                    payload_repo.add_message_reply,
-                    message_id,
-                ),
-            ),
+            }
         )
         return res
 
@@ -75,18 +105,17 @@ def send_close_order_success(chat_id: str, message_id: int):
     def _send(res: tuple[int, str]) -> tuple[int, str]:
         id, memo = res
         telegram_repo.send(
-            telegram.Channel("", chat_id),
-            compose(
-                telegram_repo.build_send_payload(
-                    payload_repo.add_close_order_success,
-                    id,
-                    memo,
+            {
+                "chat_id": chat_id,
+                "reply_to_message_id": message_id,
+                "text": "\n".join(
+                    [
+                        f"Đóng đơn hàng `{memo}` thành công `{id}`",
+                        DIVIDER,
+                        f"Check ngay: [{get_url(str(id))}]({get_url(str(id))})",
+                    ]
                 ),
-                telegram_repo.build_send_payload(
-                    payload_repo.add_message_reply,
-                    message_id,
-                ),
-            ),
+            }
         )
         return res
 
@@ -95,21 +124,22 @@ def send_close_order_success(chat_id: str, message_id: int):
 
 def send_close_order_error(chat_id: str, message_id: int):
     def _send(res: tuple[Exception, str, int]) -> tuple[Exception, str, int]:
-        err, memo, transaction_id = res
+        err, memo, id = res
         telegram_repo.send(
-            telegram.Channel("", chat_id),
-            compose(
-                telegram_repo.build_send_payload(
-                    payload_repo.add_close_order_error,
-                    err,
-                    memo,
-                    str(transaction_id),
+            {
+                "chat_id": chat_id,
+                "reply_to_message_id": message_id,
+                "text": "\n".join(
+                    [
+                        f"Đóng đơn hàng `{memo}` thất bại `{id}`",
+                        DIVIDER,
+                        f"Check ngay: [{get_url(str(id))}]({get_url(str(id))})",
+                        "```",
+                        repr(err),
+                        "```",
+                    ]
                 ),
-                telegram_repo.build_send_payload(
-                    payload_repo.add_message_reply,
-                    message_id,
-                ),
-            ),
+            }
         )
         return res
